UIImageView extension added for loading article images
