+++ FoodAllert/Controllers/Articles/ArticleTableViewCell.swift
@@ -1,0 +1,9 @@
+//
+//  ArticleTableViewCell.swift
+//  FoodAllert
+//
+//  Created by José Gil Ramírez S on 1/29/19.
+//  Copyright © 2019 José Gil Ramírez S. All rights reserved.
+//
+
+import Foundation
