+++ FoodAllert/Extensions/Extensions.swift
@@ -7,7 +7,7 @@
 //
 
 import Foundation
-
+import UIKit
 
 extension Array where Element:Equatable{
     func removeDuplicates() -> [Element] {
@@ -19,5 +19,19 @@ extension Array where Element:Equatable{
             }
         }
         return result
+    }
+}
+
+extension UIImageView {
+    func load(url: URL) {
+        DispatchQueue.global().async { [weak self] in
+            if let data = try? Data(contentsOf: url) {
+                if let image = UIImage(data: data) {
+                    DispatchQueue.main.async {
+                        self?.image = image
+                    }
+                }
+            }
+        }
     }
 }
